fix tests for unselect course
